requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
tensorflow
boto3